requirements.txt
requests>=2.28,<3
pymupdf>=1.23,<2
pdfplumber>=0.9,<1
python-dotenv>=1.0,<2
fastapi>=0.128,<1
uvicorn>=0.29,<1
anthropic>=0.40,<1
